wallpapers/image: Remove Tiled mode
We would like to improve wallpaper image memory usage. Unfortunately,
this means that we need to re-implement image item bits, including all
the various fill modes.

Unfortunately, the `Tile` mode cannot be re-implemented with vanilla
QtQuick APIs both for RHI and software rendering. However, this also
raises questions whether the tiled mode has use cases in 2025, and
whether it needs to be included with mainstream image plugin. Perhaps
it could be provided as an extra plugin to replicate a retro look.

In either case, this change removes the tiled mode in order to assist
us with our plasma optimization goals. The tiled mode plugin has
been re-implemented in work/zzag/tiled-wallpaper branch in kdeplasma-addons
repo but it has not been uploaded for review yet because we are not sure
whether the tiled mode is still being used and how much. If you have a
usecase, please let us know.